requirements.txt
click
fastapi
uvicorn
docker
PyYAML
psutil
lxc
